web_app/requirements.txt
bcrypt==4.1.2
blinker==1.9.0
cachelib==0.13.0
cffi==1.17.1
click==8.2.1
cryptography==41.0.7
Flask==3.0.0
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Session==0.5.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.2.0
packaging==25.0
protobuf==4.21.12
pycparser==2.22
pycryptodome==3.19.0
PyJWT==2.10.1
pyOpenSSL==23.3.0
python-dotenv==1.0.0
Werkzeug==3.1.3
